Some stuff for windows
