+++ getgit/cli/general.py
@@ -1,0 +1,11 @@
+def introduce_program():
+    txt = """
+    Welocme to getgit! I Hope that you will enjoy this program!
+    From people to people!
+    """
+    print(txt)
+
+def ask_git_version_service():
+    txt = "Choose git service that you use:\n"
+    gits = ("github", "gitlab")
+    return txt, gits
